feat: Update shared-types
