+++ src/models/investments/SecurityPricing.model.ts
@@ -1,3 +1,4 @@
+import { SecurityPricingModel } from 'shared-types';
 import {
   Table,
   Column,
@@ -7,18 +8,6 @@ import {
   BelongsTo,
 } from 'sequelize-typescript';
 import Security from '@models/investments/Security.model';
-
-export interface SecurityPricingAttributes {
-  // PR key
-  securityId: number;
-  // PK key
-  date: Date;
-  priceClose: string;
-  priceAsOf: Date;
-  source: string;
-  createdAt: Date;
-  updatedAt: Date;
-}
 
 /**
  * Represents the dynamic pricing information of financial securities over time.
@@ -33,7 +22,7 @@ export interface SecurityPricingAttributes {
   tableName: 'SecurityPricings',
   // indexes: [{ fields: ['securityId', 'date'], unique: true, primaryKey: true }],
 })
-export default class SecurityPricing extends Model<SecurityPricingAttributes> {
+export default class SecurityPricing extends Model<SecurityPricingModel> {
   @ForeignKey(() => Security)
   @Column({
     type: DataType.INTEGER,
